requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dj-rest-auth==6.0.0
Django
django-allauth==0.63.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-prometheus==2.3.1
django-redis==5.4.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
oauthlib==3.2.2
prometheus_client==0.20.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
drf_spectacular
drf_spectacular_sidecar
pillow
django_filter